+++ GrailDiary/Note+Markdown.swift
@@ -54,8 +54,11 @@ public extension Note {
     self.init(parsedString: buffer)
   }
 
-  static func makeBlankNote(hashtag: String? = nil) -> (Note, Int) {
+  static func makeBlankNote(title: String? = nil, hashtag: String? = nil) -> (Note, Int) {
     var initialText = "# "
+    if let title = title {
+      initialText += title
+    }
     let initialOffset = initialText.count
     initialText += "\n"
     if let hashtag = hashtag {