+++ GrailDiary/SavingTextEditViewController.swift
@@ -9,17 +9,10 @@ import UniformTypeIdentifiers
 /// Creates and wraps a TextEditViewController, then watches for changes and saves them to a database.
 /// Changes are autosaved on a periodic interval and flushed when this VC closes.
 final class SavingTextEditViewController: UIViewController, TextEditViewControllerDelegate {
-  enum ExistingOrUncreatedNote {
-    /// An unsaved note that will go into a folder when it is created.
-    case unsaved(folder: PredefinedFolder?)
-
-    /// An existing note. It lives in whatever folder it lives in.
-    case existing(identifier: Note.Identifier)
-  }
-
   /// Holds configuration settings for the view controller.
   struct Configuration {
-    var noteIdentifier: ExistingOrUncreatedNote
+    var folder: PredefinedFolder?
+    var noteIdentifier: String
     var note: Note
     var initialSelectedRange = NSRange(location: 0, length: 0)
     var autoFirstResponder = false
@@ -31,21 +24,34 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
   init(configuration: Configuration, noteStorage: NoteDatabase) {
     self.configuration = configuration
     self.noteStorage = noteStorage
-
+    self.noteIdentifier = configuration.noteIdentifier
     super.init(nibName: nil, bundle: nil)
     setTitleMarkdown(configuration.note.title)
   }
 
   /// Initializes a view controller for a new, unsaved note.
-  convenience init(database: NoteDatabase, folder: PredefinedFolder?, currentHashtag: String? = nil, autoFirstResponder: Bool = false) {
-    let (note, initialOffset) = Note.makeBlankNote(hashtag: currentHashtag)
+  convenience init(
+    database: NoteDatabase,
+    folder: PredefinedFolder?,
+    title: String? = nil,
+    initialImage: UIImage? = nil,
+    currentHashtag: String? = nil,
+    autoFirstResponder: Bool = false
+  ) {
+    let (note, initialOffset) = Note.makeBlankNote(title: title, hashtag: currentHashtag)
     let configuration = Configuration(
-      noteIdentifier: .unsaved(folder: folder),
+      folder: folder,
+      noteIdentifier: UUID().uuidString,
       note: note,
       initialSelectedRange: NSRange(location: initialOffset, length: 0),
       autoFirstResponder: autoFirstResponder
     )
     self.init(configuration: configuration, noteStorage: database)
+    if let initialImage = initialImage,
+       let convertedData = initialImage.jpegData(compressionQuality: 0.8)
+    {
+      insertImageData(convertedData, type: .jpeg)
+    }
   }
 
   @available(*, unavailable)
@@ -76,12 +82,7 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
   private var autosaveTimer: Timer?
 
   /// The identifier for the displayed note; nil if the note is not yet saved to the database.
-  var noteIdentifier: Note.Identifier? {
-    guard case .existing(let noteIdentifier) = configuration.noteIdentifier else {
-      return nil
-    }
-    return noteIdentifier
-  }
+  let noteIdentifier: Note.Identifier
 
   /// The current note.
   var note: Note { configuration.note }
@@ -135,9 +136,11 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
       if splitViewController?.isCollapsed ?? false {
         navigationItem.rightBarButtonItem = nil
         navigationController?.isToolbarHidden = false
-        toolbarItems = [UIBarButtonItem.flexibleSpace(), AppCommandsButtonItems.newNote()]
+        if let newNoteButton = toolbarButtonBuilder?.makeNewNoteButtonItem() {
+          toolbarItems = [UIBarButtonItem.flexibleSpace(), newNoteButton]
+        }
       } else {
-        navigationItem.rightBarButtonItem = AppCommandsButtonItems.newNote()
+        navigationItem.rightBarButtonItem = toolbarButtonBuilder?.makeNewNoteButtonItem()
         navigationController?.isToolbarHidden = true
         toolbarItems = []
       }
@@ -145,30 +148,20 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
   }
 
   /// Writes a note to storage.
-  @discardableResult
-  private func updateNote(_ note: Note) throws -> Note.Identifier {
+  private func updateNote(_ note: Note) throws {
     assert(Thread.isMainThread)
     var note = note
     // Copy over the initial reference, if any
     note.reference = configuration.note.reference
     setTitleMarkdown(note.title)
-    switch configuration.noteIdentifier {
-    case .existing(identifier: let noteIdentifier):
-      Logger.shared.info("SavingTextEditViewController: Updating note \(noteIdentifier)")
-      try noteStorage.updateNote(noteIdentifier: noteIdentifier, updateBlock: { oldNote in
-        var mergedNote = note
-        mergedNote.copyContentKeysForMatchingContent(from: oldNote)
-        mergedNote.folder = oldNote.folder
-        return mergedNote
-      })
-      return noteIdentifier
-    case .unsaved(folder: let folder):
-      note.folder = folder?.rawValue
-      let noteIdentifier = try noteStorage.createNote(note)
-      Logger.shared.info("SavingTextEditViewController: Created note \(noteIdentifier)")
-      configuration.noteIdentifier = .existing(identifier: noteIdentifier)
-      return noteIdentifier
-    }
+    note.folder = configuration.folder?.rawValue
+    Logger.shared.info("SavingTextEditViewController: Updating note \(noteIdentifier)")
+    try noteStorage.updateNote(noteIdentifier: noteIdentifier, updateBlock: { oldNote in
+      var mergedNote = note
+      mergedNote.copyContentKeysForMatchingContent(from: oldNote)
+      mergedNote.folder = oldNote.folder
+      return mergedNote
+    })
   }
 
   private func tryUpdateNote(_ note: Note) {
@@ -194,6 +187,30 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
     tryUpdateNote(note)
     hasUnsavedChanges = false
     completion?()
+  }
+
+  /// Saves the current contents to the database, whether or not hasUnsavedChanges is true.
+  private func forceSave() throws {
+    let note = Note(parsedString: textEditViewController.textStorage.storage.rawString)
+    try updateNote(note)
+    hasUnsavedChanges = false
+  }
+
+  /// Inserts an image at the current insertion point.
+  private func insertImageData(_ imageData: Data, type: UTType) {
+    do {
+      try forceSave()
+      let reference = try storeImageData(imageData, type: type)
+      let markdown = "\n\n![](\(reference))\n\n"
+      let initialRange = textEditViewController.selectedRange
+      var rawRange = textEditViewController.textStorage.storage.rawStringRange(forRange: initialRange)
+      rawRange.location += markdown.utf16.count
+      rawRange.length = 0
+      textEditViewController.textStorage.replaceCharacters(in: initialRange, with: markdown)
+      textEditViewController.selectedRange = textEditViewController.textStorage.storage.range(forRawStringRange: rawRange)
+    } catch {
+      Logger.shared.error("Could not save initial image: \(error)")
+    }
   }
 
   func textEditViewControllerDidChangeContents(_ viewController: TextEditViewController) {
@@ -223,20 +240,10 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
 
 extension SavingTextEditViewController: ImageStorage {
   func storeImageData(_ imageData: Data, type: UTType) throws -> String {
-    let noteIdentifier = try ensureNoteIdentifier()
     return try noteStorage.writeAssociatedData(imageData, noteIdentifier: noteIdentifier, role: "embeddedImage", type: type)
   }
 
-  private func ensureNoteIdentifier() throws -> Note.Identifier {
-    if let noteIdentifier = noteIdentifier {
-      return noteIdentifier
-    } else {
-      return try updateNote(Note(parsedString: textEditViewController.textStorage.storage.rawString))
-    }
-  }
-
   func retrieveImageDataForKey(_ key: String) throws -> Data {
-    let noteIdentifier = try ensureNoteIdentifier()
     return try noteStorage.readAssociatedData(from: noteIdentifier, key: key)
   }
 }