+++ GrailDiary/WebViewController.swift
@@ -67,9 +67,9 @@ public final class WebViewController: UIViewController, ReferenceViewController 
     if splitViewController?.isCollapsed ?? false {
       navigationItem.rightBarButtonItems = []
       navigationController?.isToolbarHidden = false
-      toolbarItems = [showNotesButton, UIBarButtonItem.flexibleSpace(), AppCommandsButtonItems.newNote()]
+      toolbarItems = [showNotesButton, UIBarButtonItem.flexibleSpace(), toolbarButtonBuilder?.makeNewNoteButtonItem()].compactMap { $0 }
     } else {
-      navigationItem.rightBarButtonItems = [AppCommandsButtonItems.newNote(), showNotesButton]
+      navigationItem.rightBarButtonItems = [toolbarButtonBuilder?.makeNewNoteButtonItem(), showNotesButton].compactMap { $0 }
       navigationController?.isToolbarHidden = true
       toolbarItems = []
     }