+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Google Books integration! You can either add book details to an existing note or create a new note based upon a book.
+
 ## [0.42.0] - 2020-02-13
 
 ### Added